fix some bug

+++ app/src/main/java/com/oggysocial/oggysocial/adapters/CommentTouchHelper.java
@@ -7,7 +7,11 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.firebase.auth.FirebaseAuth;
 import com.oggysocial.oggysocial.R;
+import com.oggysocial.oggysocial.models.Comment;
+
+import java.util.Objects;
 
 import it.xabaras.android.recyclerview.swipedecorator.RecyclerViewSwipeDecorator;
 
@@ -19,11 +23,12 @@ public class CommentTouchHelper extends ItemTouchHelper.SimpleCallback {
 
     @Override
     public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
-        return false;
+        return true;
     }
 
     @Override
     public void onChildDraw(@NonNull Canvas c, @NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
+
         new RecyclerViewSwipeDecorator.Builder(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive)
                 .addBackgroundColor(ContextCompat.getColor(recyclerView.getContext(), R.color.bg_delete))
                 .addSwipeLeftActionIcon(R.drawable.trash_line)
@@ -31,15 +36,30 @@ public class CommentTouchHelper extends ItemTouchHelper.SimpleCallback {
                 .create()
                 .decorate();
 
+
         super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
     }
 
     @Override
+    public int getSwipeDirs(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder) {
+        CommentAdapter adapter = (CommentAdapter) viewHolder.getBindingAdapter();
+        assert adapter != null;
+        Comment comment = adapter.comments.get(viewHolder.getAbsoluteAdapterPosition());
+
+        if (!comment.getAuthor().getId().equals(FirebaseAuth.getInstance().getUid())) {
+            return 0;
+        }
+
+        return super.getSwipeDirs(recyclerView, viewHolder);
+    }
+
+    @Override
     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-        if (direction == ItemTouchHelper.LEFT) {
+        CommentAdapter adapter = (CommentAdapter) viewHolder.getBindingAdapter();
+        assert adapter != null;
+        Comment comment = adapter.comments.get(viewHolder.getAbsoluteAdapterPosition());
+        if (direction == ItemTouchHelper.LEFT && Objects.equals(comment.getAuthor().getId(), FirebaseAuth.getInstance().getUid())) {
             int position = viewHolder.getAbsoluteAdapterPosition();
-            CommentAdapter adapter = (CommentAdapter) viewHolder.getBindingAdapter();
-            assert adapter != null;
             adapter.deleteComment(position);
         }
     }
